// testing/types/user.d.ts
export interface paths {
    "/user/profile": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get profile
         * @description Gets the user profile for the current session
         */
        get: operations["get_profile"];
        /**
         * Update profile
         * @description Updates the user profile for the current session
         */
        put: operations["update_profile"];
        /**
         * Create profile
         * @description Creates a new user profile for the current session
         */
        post: operations["create_profile"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/user/profile/avatar": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        /**
         * Replace profile avatar
         * @description Returns the URL that should be used to upload an image of the user's new avatar
         */
        put: operations["put_avatar"];
        post?: never;
        /**
         * Delete profile avatar
         * @description Delete's the user profile's avatar from wherever it is stored
         */
        delete: operations["delete_avatar"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}