// testing/types/timekeeper.d.ts
export interface paths {
    "/timekeeper/tag": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Retreive tags
         * @description Retreives all the tags associated with the user, as well as the default ones
         */
        get: operations["get_tags"];
        put?: never;
        /**
         * Create a tag
         * @description Creates a new tag in the database. The returned tag ID can be used in other endpoints in this service.
         */
        post: operations["create_tag"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/timekeeper/tag/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        /**
         * Modify tag
         * @description Modifies the fields of the tag at the ID
         */
        put: operations["modify_tag"];
        post?: never;
        /**
         * Delete a tag
         * @description Deletes the tag at the given ID. This just marks the tag as deleted, and doesn't actually delete the tag in the database. Timers posted with a deleted tag will still have that tag, the tag just won't appear when querying some endpoints.
         */
        delete: operations["delete_tag"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/timekeeper/time-split": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Retreive time splits
         * @description Retrieves all the time splits associated with the user, as well as the default ones
         */
        get: operations["get_time_splits"];
        put?: never;
        /**
         * Create a time split
         * @description Creates a new time split in the database. The returned time split ID can be used in other endpoints in this service. The timer lengths should use the `jiff::Span` format
         */
        post: operations["create_time_split"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/timekeeper/time-split/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        /**
         * Modify a time split
         * @description Modifies the fields of the time split at the ID.
         */
        put: operations["modify_time_split"];
        post?: never;
        /**
         * Delete a time split
         * @description Deletes the time split at the given ID. This just marks the time split as deleted, and doesn't actually delete the time split in the database. Timers posted with a deleted time split will still have that time split, the time split just won't appear when querying some endpoints.
         */
        delete: operations["delete_time_split"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/timekeeper/timer": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Retreive a timer
         * @description Retreives the most recently started/ended timer if no parameters are specified. Otherwise returns the timers that match the parameters.
         */
        get: operations["get_timers"];
        /**
         * Modify a timer
         * @description This route replaces the fields of the timer at the passed in start time with new fields. When a timer is paused, this route should be used to change the end_timestamp of the timer to the Unix timestamp at which the timer was paused. To resume a timer, a new timer should be posted with the Unix timestamp at which the timer was resumed, and the Unix timestamp in the future at which the remaining time will have elapsed.
         */
        put: operations["modify_timer"];
        /**
         * Post a timer
         * @description Upon starting a new timer, the Unix timestamp of when the timer was started, as well as the Unix timestamp in the future when the timer will end should be sent to this service. The ID of the time split, whether the timer is a work or a break timer, and the tag IDs associated with the timer should also be sent.
         */
        post: operations["post_timer"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}